+++ python/src/squishy_volumes_extension/panels/panel_input.py
@@ -21,7 +21,7 @@ import bpy
 from typing import Any
 
 from ..get_preferences import get_confirm_bake_overwrite
-from ..nodes.drivers import add_drivers
+from ..drivers import add_drivers
 
 from ..squishy_volumes_properties import (
     get_selected_input_object,
